Move transformation of rules T array to Analyser interface func
Go does not allow the usage of an interface on a typed array. To
resolve this issue we manually create a new array with the interface type
and add all items to it. This no happens inside the rules sturct
instead of being a single standalone function inside the main package.